+++ st2client/st2client/shell.py
@@ -47,6 +47,7 @@ from st2client.commands import timer
 from st2client.commands import webhook
 from st2client.commands import rule
 from st2client.commands import rule_enforcement
+from st2client.commands import rbac
 from st2client.config import set_config
 from st2client.exceptions.operations import OperationFailureException
 from st2client.utils.logging import LogLevelFilter, set_log_level_for_all_loggers
@@ -252,6 +253,14 @@ class Shell(BaseCLIApp):
 
         self.commands['rule-enforcement'] = rule_enforcement.RuleEnforcementBranch(
             'Models that represent enforcement of rules.',
+            self, self.subparsers)
+
+        # RBAC
+        self.commands['role'] = rbac.RoleBranch(
+            'RBAC roles.',
+            self, self.subparsers)
+        self.commands['role-assignment'] = rbac.RoleAssignmentBranch(
+            'RBAC role assignments.',
             self, self.subparsers)
 
     def run(self, argv):